requirements.txt
feedparser>=6.0.10
httpx>=0.24.0
trafilatura>=1.6.0
deep-translator>=1.10.1
python-dotenv>=1.0.0
